Extract message producers into their own classes
This takes the quickly-growing pattern match in `Producer.apply`, and
splits it into individual classes under the `sectery.producers` package.

Adding new producers now consists of writing a new `Producer`
implementation, and then adding it to the list of producers in
`Producer.producers`.
