+++ src/main/scala/sectery/MessageQueues.scala
@@ -1,14 +1,7 @@
 package sectery
 
-import java.io.IOException
-import java.text.SimpleDateFormat
-import java.util.concurrent.TimeUnit
-import java.util.Date
-import java.util.Locale
-import java.util.TimeZone
 import zio.clock.Clock
 import zio.duration._
-import zio.Fiber
 import zio.Queue
 import zio.Schedule
 import zio.UIO
@@ -25,26 +18,6 @@ case class Rx(channel: String, nick: String, message: String)
  * A message to send to IRC.
  */
 case class Tx(channel: String, message: String)
-
-/**
- * Read incoming messages, and produce any number of responses.
- */
-object Producer {
-  def apply(m: Rx): URIO[Clock, Iterable[Tx]] =
-    m match
-      case Rx(c, _, "@ping") =>
-        ZIO.effectTotal(Some(Tx(c, "pong")))
-      case Rx(c, _, "@time") =>
-        for
-          millis <- zio.clock.currentTime(TimeUnit.MILLISECONDS)
-          date    = new Date(millis)
-          sdf     = new SimpleDateFormat("EEE, d MMM yyyy, kk:mm zzz")
-          _       = sdf.setTimeZone(TimeZone.getTimeZone("America/Phoenix"))
-          pretty  = sdf.format(date)
-        yield Some(Tx(c, pretty))
-      case _ =>
-        ZIO.effectTotal(None)
-}
 
 /**
  * An interface for defining how to send an outgoing message, which
